+++ oneview_redfish_toolkit/tests/test_util.py
@@ -188,8 +188,8 @@ class TestUtil(unittest.TestCase):
         except Exception as e:
             self.fail('Failed to load registries files: {}'.format(e.msg))
 
-    @mock.patch.object(util, 'OneViewClient')
-    def test_load_config(self, mock_ov):
+    @mock.patch.object(util, 'check_oneview_availability')
+    def test_load_config(self, check_ov_availability):
         # Test load config function
 
         util.load_config(self.config_file)
@@ -199,7 +199,7 @@ class TestUtil(unittest.TestCase):
         self.assertIsNotNone(util.ov_config, msg='Failed do create ov_config')
         self.assertIsNotNone(
             util.registry_dict, msg='Failed to load registries')
-        self.assertIsNotNone(util.ov_client, msg='Failed to connect to OV')
+        self.assertTrue(check_ov_availability.called)
 
     @mock.patch.object(util, 'OneViewClient')
     def test_get_ov_client_recover(self, oneview_client_mockup):
@@ -253,8 +253,9 @@ class TestUtil(unittest.TestCase):
         except Exception:
             self.fail("Failed to get a valid IP Address")
 
-    @mock.patch.object(util, 'OneViewClient')
-    def test_create_certs(self, oneview_client_mockup):
+    @mock.patch.object(util, 'check_oneview_availability')
+    def test_create_certs(
+        self, check_ov_availability):
         # Test generate_certificate function
 
         util.load_config(self.config_file)
@@ -264,19 +265,18 @@ class TestUtil(unittest.TestCase):
         self.assertTrue(os.path.exists(os.path.join("certs", "test" + ".crt")))
         self.assertTrue(os.path.exists(os.path.join("certs", "test" + ".key")))
 
-    @mock.patch.object(util, 'OneViewClient')
-    @mock.patch('oneview_redfish_toolkit.util.delivery_retry_attempts', '')
-    @mock.patch('oneview_redfish_toolkit.util.delivery_retry_interval', '')
-    def test_load_event_service_invalid_info(self, oneview_client_mockup):
+    @mock.patch.object(util, 'check_oneview_availability')
+    def test_load_event_service_invalid_info(
+        self, check_ov_availability):
         self.assertRaises(
             OneViewRedfishError, util.load_config(self.config_file))
 
-    @mock.patch.object(util, 'OneViewClient')
+    @mock.patch.object(util, 'check_oneview_availability')
     @mock.patch.object(util, 'subscriptions_by_type')
     @mock.patch(
         'oneview_redfish_toolkit.event_dispatcher.EventDispatcher.start')
     def test_submit_event_with_subscriber(
-        self, start_mock, subscription_mock, ov_mock):
+        self, start_mock, subscription_mock, check_ov_availability):
         """Tests SubmitTestEvent action with two subscribers"""
 
         util.load_config(self.config_file)
@@ -295,12 +295,12 @@ class TestUtil(unittest.TestCase):
 
         self.assertTrue(start_mock.call_count == 2)
 
-    @mock.patch.object(util, 'OneViewClient')
+    @mock.patch.object(util, 'check_oneview_availability')
     @mock.patch.object(util, 'subscriptions_by_type')
     @mock.patch(
         'oneview_redfish_toolkit.event_dispatcher.EventDispatcher.start')
     def test_submit_event_without_subscriber(
-        self, start_mock, subscription_mock, ov_mock):
+        self, start_mock, subscription_mock, check_ov_availability):
         """Tests SubmitTestEvent action with no subscribers"""
 
         util.load_config(self.config_file)